Refactor imports and remove unused sidebar configuration files; add NotFound component and health check API route.
