+++ src/app/(root)/(dashboard)/meeting/[id]/page.tsx
@@ -8,8 +8,7 @@ import { Loader } from "lucide-react";
 
 import { useGetCallById } from "@/hooks/useGetCallById";
 import Alert from "@/components/Alert";
-import MeetingSetup from "@/components/MeetingSetup";
-import MeetingRoom from "@/components/MeetingRoom";
+import { MeetingRoom, MeetingSetup } from "@/components/workspace/meeting";
 
 const MeetingPage = () => {
   const { id } = useParams();
