--- src/constants/sidebar.json
@@ -1,52 +1,0 @@
-[
-  {
-    "id": "workspace-alpha",
-    "name": "Workspace Alpha",
-    "route": "/dashboard/workspace/workspace-alpha"
-  },
-  {
-    "id": "workspace-beta",
-    "name": "Workspace Beta",
-    "route": "/dashboard/workspace/workspace-beta"
-  },
-  {
-    "id": "workspace-gamma",
-    "name": "Workspace Gamma",
-    "route": "/dashboard/workspace/workspace-gamma"
-  },
-  {
-    "id": "workspace-delta",
-    "name": "Workspace Delta",
-    "route": "/dashboard/workspace/workspace-delta"
-  },
-  {
-    "id": "workspace-epsilon",
-    "name": "Workspace Epsilon",
-    "route": "/dashboard/workspace/workspace-epsilon"
-  },
-  {
-    "id": "workspace-zeta",
-    "name": "Workspace Zeta",
-    "route": "/dashboard/workspace/workspace-zeta"
-  },
-  {
-    "id": "workspace-eta",
-    "name": "Workspace Eta",
-    "route": "/dashboard/workspace/workspace-eta"
-  },
-  {
-    "id": "workspace-theta",
-    "name": "Workspace Theta",
-    "route": "/dashboard/workspace/workspace-theta"
-  },
-  {
-    "id": "workspace-iota",
-    "name": "Workspace Iota",
-    "route": "/dashboard/workspace/workspace-iota"
-  },
-  {
-    "id": "workspace-kappa",
-    "name": "Workspace Kappa",
-    "route": "/dashboard/workspace/workspace-kappa"
-  }
-]
